Reorganize architecture dependant files
