+++ kernel/arch/x86_64/regs.rs
@@ -1,22 +1,5 @@
-#[cfg(target_arch = "x86")]
 #[derive(Copy, Clone, Default)]
-pub struct Regs {
-    pub ax: usize,
-    pub bx: usize,
-    pub cx: usize,
-    pub dx: usize,
-    pub di: usize,
-    pub si: usize,
-    pub bp: usize,
-    pub ip: usize,
-    pub cs: usize,
-    pub flags: usize,
-    pub sp: usize,
-    pub ss: usize,
-}
-
-#[cfg(target_arch = "x86_64")]
-#[derive(Copy, Clone, Default)]
+#[repr(packed)]
 pub struct Regs {
     pub ax: usize,
     pub bx: usize,
