--- kernel/scheduler/mod.rs
@@ -1,8 +1,0 @@
-pub use self::context::Context;
-pub use self::regs::Regs;
-pub use self::tss::TSS;
-
-/// Context
-pub mod context;
-pub mod regs;
-pub mod tss;
