+++ kernel/sync/rwlock.rs
@@ -3,7 +3,7 @@ use core::ops::{Deref, DerefMut};
 
 use sync::Mutex;
 
-use scheduler::context::context_switch;
+use arch::context::context_switch;
 
 struct RwLockInner {
     writer: bool,
