+++ kernel/usb/ohci.rs
@@ -8,7 +8,7 @@ use core::mem;
 use drivers::io::{Io, Mmio};
 use drivers::pci::config::PciConfig;
 
-use scheduler::context::context_switch;
+use arch::context::context_switch;
 
 use schemes::KScheme;
